03_journey_into_microservices/demo/backend/auth_microservice/requirements.txt
Django==3.2
redis==4.3.1
celery==4.4.0
django-celery-beat==1.4.0
djangorestframework==3.11.0
djangorestframework-sso==0.3.1
django-redis==4.11.0
python-decouple==3.3
mysqlclient==1.4.6
graphene==2.1.8
graphene-django==2.13.0
djangorestframework-simplejwt==5.3.0
cryptography==2.9.2